refactor: controllerTest -> documentationTest로 병합
