+++ backend/src/test/java/com/woowacourse/gongcheck/documentation/JobDocumentation.java
@@ -12,30 +12,35 @@ import com.woowacourse.gongcheck.application.response.JobsResponse;
 import com.woowacourse.gongcheck.domain.host.Host;
 import com.woowacourse.gongcheck.domain.space.Space;
 import java.util.List;
+import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 
 class JobDocumentation extends DocumentationTest {
 
-    @Test
-    void 작업_조회() {
-        Host host = Host_생성("1234");
-        Space space = Space_생성(host, "잠실");
-        when(jobService.findPage(anyLong(), anyLong(), any())).thenReturn(
-                JobsResponse.of(List.of(
-                                Job_생성(space, "청소"),
-                                Job_생성(space, "마감")),
-                        true)
-        );
-        when(authenticationContext.getPrincipal()).thenReturn(String.valueOf(anyLong()));
+    @Nested
+    class 작업을_조회한다 {
 
-        docsGiven
-                .header("Authorization", "Bearer jwt.token.here")
-                .queryParam("page", 0)
-                .queryParam("size", 2)
-                .when().get("/api/spaces/1/jobs")
-                .then().log().all()
-                .apply(document("jobs/list"))
-                .statusCode(HttpStatus.OK.value());
+        @Test
+        void 작업_조회에_성공한다() {
+            Host host = Host_생성("1234");
+            Space space = Space_생성(host, "잠실");
+            when(jobService.findPage(anyLong(), anyLong(), any())).thenReturn(
+                    JobsResponse.of(List.of(
+                                    Job_생성(space, "청소"),
+                                    Job_생성(space, "마감")),
+                            true)
+            );
+            when(authenticationContext.getPrincipal()).thenReturn(String.valueOf(anyLong()));
+
+            docsGiven
+                    .header("Authorization", "Bearer jwt.token.here")
+                    .queryParam("page", 0)
+                    .queryParam("size", 2)
+                    .when().get("/api/spaces/1/jobs")
+                    .then().log().all()
+                    .apply(document("jobs/list"))
+                    .statusCode(HttpStatus.OK.value());
+        }
     }
 }
